Add support for negative sizes
This requires changing the underlying type for size extension (on numeric
operations) and for the default constants from `u64` to `i64`. This is a
breaking change but all uses encapsulated in a `Size` type should be silently
upgraded to the new underlying type without issue.

This also introduces support for subtracting sizes from each other.