Reduce branchiness at the cost of (internal) type divergence
The formatting rules are only ever called on the absolute value of an underlying
size, so we can use `u64` there instead of `i64`. Then, because the maximum
value of an underlying `i64` is less than that of the maximum value of the
formatter rule `u64`, we can avoid the binary search result branch for handling
a case where the found formatter is larger than the last rule in the array
(because a `Size` formatted via an intermediate `i64` can never be larger than
`u64::max_value()`).