Introduce a separate `Size::from_const()` function
This is in lieu of having `Size::from_bytes()` be a `const` function, which came
at the heavy cost of preventing is likely be the most-used function in the crate
from being generic.

Importantly, this makes it possible to now do the following:

```rust
use std::fs::*;

let metadata = File::metadata("foo").unwrap();
let file_size = Size::from_bytes(metadata.len());
```

without needing to use `metadata.len() as i64` (because it `Metadata::len()`
returns a `u64`).