Simplify converting `Size` to `String`
Previously, a `format!("{}", size)` call or else a call to
`size.to_string(Base::Base2, Style::Smart)` with all the parameters was required
to obtain a `String` from a `Size`.

Now `Size::to_string()` returns a `String` and requires no parameters, but at
the cost of moving formatting to a separate type (a `SizeFormatter` returned by
`Size::format()`).