Coalesce FormattableSize and SizeFormatter types
We weren't able to use a trait for the reasons previously explained (mainly the
need for the user to `use` the trait) but this should work. We take advantage of
the fact that rust can infer the generic type based off which impl offers a
function `SizeFormatter<?>::new()` which lets users omit the `<()>` from the
expression when instantiating a new `SizeFormatter`, and then we use a specific
generic type to provide a `Display` impl and the `to_string()` method.

This should not be considered a breaking change and all the old tests pass
without modification and without running into any type ambiguity.