Reduce unneeded trim operations in Size::from_str()
We already trim the input at the very start (so we can read only alphabetical
characters from the end) and stop at the first alphabetical char, so there's no
need to trim `unit`.

For `number`, the start has already been trimmed so we only need to trim any
spaces captured between the scalar quantity and the start of the unit name.